Move core asset upload code to files.py
